Move App and MainRouter to project root
