+++ src/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import MainRouter from './modules/MainRouter'
+import MainRouter from './MainRouter'
 
 // eslint-disable-next-line no-undef
 ReactDOM.render(<MainRouter />, document.getElementById('app-root'))
